# src/main/java/com/example/bankcards/dto/UserResponse.java
package com.example.bankcards.dto;

import com.example.bankcards.enums.Role;
import io.swagger.v3.oas.annotations.media.Schema;

@Schema(description = "User response DTO containing user details")
public class UserResponse {

    @Schema(description = "Unique identifier of the user", example = "1")
    private Long id;

    @Schema(description = "Username of the user", example = "john_doe")
    private String username;

    @Schema(description = "Role of the user", example = "USER")
    private Role role;

    public Long getId() {
        return id;
    }

    public void setId(Long id) {
        this.id = id;
    }

    public String getUsername() {
        return username;
    }

    public void setUsername(String username) {
        this.username = username;
    }

    public Role getRole() {
        return role;
    }

    public void setRole(Role role) {
        this.role = role;
    }
}